% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAudition.R
\name{addAudition}
\alias{addAudition}
\title{addAudition}
\usage{
addAudition(seqs, threshold, include_ncbi = F, just_ID = F,
  make_blast = F, python_path = "/usr/local/bin/python3",
  validate_name = F)
}
\arguments{
\item{seqs}{sequence file}

\item{threshold}{threshold}

\item{include_ncbi}{include_ncbi}

\item{just_ID}{just_ID}

\item{make_blast}{make_blast}

\item{python_path}{python_path}

\item{validate_name}{validate_name}
}
\value{
species ID plus DNA barcode auditing
}
\description{
This function adds an audition step
(Oliveira et al 2016, doi: 10.1111/jfb.13169)
to each selected specimen by \code{\link{ID_engine}}, given a certain threshold.
This function, in turn, uses another function called `AuditionBarcodes()`.
This prior function is coupled with addAudition and can validate species names by
taking accepted names from Worms database.
}
