% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAudition.R
\name{auditOnID}
\alias{auditOnID}
\title{auditOnID}
\usage{
auditOnID(seqs, threshold = 0.99, exclude_ncbi = TRUE,
  just_ID = FALSE, make_blast = FALSE, validate_name = FALSE)
}
\arguments{
\item{seqs}{sequence file}

\item{threshold}{minimal similarity proportion allowed between sample and specimen matches}

\item{exclude_ncbi}{exclude barcodes obtained from NCBI when DNA barcode auditing is allowed, i.e. not just_ID}

\item{just_ID}{if TRUE, audition step is skipped}

\item{make_blast}{if TRUE, blastn algorithm is run for matching sequences}

\item{validate_name}{if TRUE, validation of species name is performed. This validation of name is performed
by using taxamatch algorithm from \href{WoRMS Rest API}{http://www.marinespecies.org/aphia.php?p=webservice}}
}
\value{
species ID plus DNA barcode auditing
}
\description{
This function adds an audition step
(\href{doi: 10.1111/jfb.13169}{Oliveira et al 2016})
to each selected specimen by \code{\link{ID_engine}}, given a certain threshold.
This function, in turn, uses another function called \code{\link{AuditionBarcodes}}.
This prior function is coupled with addAudition and can validate species names by
taking accepted names from \href{http://www.marinespecies.org/index.php}{Worms database}.
}
\examples{
\dontrun{
seqs  = system.file("sequences.fa", package = "boldminer")
boldminer::auditOnID(seqs, just_ID = TRUE)

}
}
