% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getIDfromBOLD}
\alias{getIDfromBOLD}
\title{getIDfromBOLD}
\usage{
getIDfromBOLD(seq, db)
}
\arguments{
\item{seq}{a string containing a dna sequence}

\item{db}{database available for BOLD's API}
}
\value{
data frame with ID
}
\description{
getIDfromBOLD
}
