% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID_engine.R
\name{ID_engine}
\alias{ID_engine}
\title{ID_engine}
\usage{
ID_engine(query, db, make_blast = TRUE, ...)
}
\arguments{
\item{query}{query}

\item{db}{db}

\item{make_blast}{make_blast}

\item{...}{amazing three points}
}
\value{
species ID
}
\description{
ID_engine finds best matches between a query sequence and a
database of BOLD by using BLASTn-based algorithms.
Arguments of this function are `query` and `db`. The first
one are query sequences and the second one are one
of avilable databases in BOLD:

- `COX1`

- `COX1_SPECIES`

- `COX1_SPECIES_PUBLIC`

- `COX1_L640bp`
}
\examples{
\dontrun{
fasta_file <- system.file("sequences.fa", package = "boldminer")
out <- boldminer::ID_engine(query = fasta_file, db = "COX1_SPECIES")

}

}
