% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID_engine.R
\name{lookID}
\alias{lookID}
\title{lookID}
\usage{
lookID(out, first = 5, last = 0)
}
\arguments{
\item{out}{list of results from \code{\link{ID_engine}} function}

\item{first}{number of first rows from input}

\item{last}{number of last rows from input}
}
\value{
data frame with constrained results
}
\description{
Take first and last rows from \code{\link{ID_engine}} function results
}
\examples{
\dontrun{

boldminer::lookID(out)
}

}
