% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenData.R
\name{SpecimenData}
\alias{SpecimenData}
\title{SpecimenData}
\usage{
SpecimenData(taxon, ids, bin, container, institutions, researchers, geo,
  ...)
}
\arguments{
\item{taxon}{Species or taxa to look for}

\item{ids}{IDs from specimen records. e.g. ANGBF12704-15}

\item{bin}{barcode-based cluster ID. e.g. BOLD:AAA4689}

\item{container}{container name. e.g. FIPP}

\item{geo}{name of geographical area to look for.}

\item{...}{others query name or column name that can be present in metadata}

\item{institution}{institution names. e.g. Smithsonian Institution}

\item{reseachers}{names of identifiers or specimen collectors}
}
\value{
metadata information
}
\description{
This function let us mine associated metadata
from any specimen according to query arguments.
It takes very specific queries and
use them as headers for BOLD API host
for downloading metadata. More than one query
can be specified by using "|".
}
