% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenData.R
\name{sumSData}
\alias{sumSData}
\title{sumSData}
\usage{
sumSData(df, cols)
}
\arguments{
\item{df}{specimen data from \code{\link{SpecimenData}} function}

\item{columns}{columns to summarize on}
}
\value{
summarized data frame
}
\description{
short utility to summarize data frame
}
\examples{
\dontrun{
specimendata <- boldminer::SpecimenData(taxon = "Elasmobranchii", geo = "Peru")
boldminer::sumSData(df = specimendata, cols = c("species_name", "country"))
}
}
