% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAudition.R
\name{addAudition}
\alias{addAudition}
\title{addAudition}
\usage{
addAudition(seqs, threshold, exlclude_ncbi = TRUE, just_ID = FALSE,
  make_blast = FALSE, validate_name = FALSE)
}
\arguments{
\item{seqs}{sequence file}

\item{threshold}{threshold}

\item{just_ID}{just_ID}

\item{make_blast}{make_blast}

\item{validate_name}{validate_name}

\item{exclude_ncbi}{exclude barcodes obtained from NCBI when DNA barcode auditing is allowed, i.e. not just_ID}
}
\value{
species ID plus DNA barcode auditing
}
\description{
This function adds an audition step
(Oliveira et al 2016, doi: 10.1111/jfb.13169)
to each selected specimen by \code{\link{ID_engine}}, given a certain threshold.
This function, in turn, uses another function called `AuditionBarcodes()`.
This prior function is coupled with addAudition and can validate species names by
taking accepted names from Worms database.
}
