% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuditionBarcodes.R
\name{AuditionBarcodes}
\alias{AuditionBarcodes}
\title{AuditionBarcodes}
\usage{
AuditionBarcodes(species, matches = NULL, validate_name = FALSE,
  exclude_ncbi = TRUE, quiet = FALSE)
}
\arguments{
\item{species}{species}

\item{matches}{matches}

\item{validate_name}{validate_name}

\item{exclude_ncbi}{exclude barcodes obtained from NCBI}

\item{quiet}{let to use running messages while auditing}
}
\value{
DNA barcode auditing
}
\description{
Despite AuditionBarcodes function is coupled with \code{\link{addAudition}} function,
it can also work with just a list of names. Furthermore, there is an argument
which enables to chose if sequences from GenBank are considered.
It is pending, however, assess whether these sequences used to assess barcode's quality
come from either a published article or direct submission.
Please notice that grades are obtained with accepted names of
species according to WoRMS database by
using the worms.py script.
Hence, since currently accepted names within `species` vector has not been
figured out, unevenness between the column `BIN_structure` and `species` could pop up.
For this reason, on above example, _Morula margariticola_ has another name into BIN structure
column (i.e. _Drupella margariticola_) which actually is its currently accepted name.
}
\examples{
\dontrun{
species <- c( "Caranx ruber", "Bathygobius lineatus", "Diodon hystrix")
boldminer::AuditionBarcodes(species, exclude_ncbi = FALSE, validate_name = TRUE)
}
}
